requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
groq>=0.4.0
numpy>=1.24.0